Keep actions/ to operations whose result is persisted
An action is an operation whose result is stored — a mutation, or a
computation whose output is written to the batch. Three files in actions/
computed values that only ever existed live.

The app CLAUDE.md is part of why. Its "Derived batch data" section used
one word for batch.shopping (computed AND stored) and for the
schedule/actuals/progress views (computed AND discarded), so "derived
batch data" read as though it covered both and the folder drifted along
exactly that seam.

Moved to model/, beside the types they read — not a new convention, since
model/brewable.ts has held phaseLabel, canRemovePhase and defaultBrewable
all along:

  brewTimer.ts          -> model/timer.ts, where TimerEvent is defined
  brewableResources.ts  -> model/brewable.ts, where Assignment is defined
  batchProgress.ts      -> model/batchProgress.ts, its own file, because it
                           spans Batch, BrewablePhase and the tracker

The section is now two: Actions, for anything persisted, and Live
computation, for what is thrown away. Both state the membership test, and
the second says plainly not to call live values "derived".

Two comments on indexedResourcesOf were corrected on the way, since they
would have been wrong in their new home: they named deriveSchedule
(renamed useSchedule in #564) and described the write-through dot-paths
into brewable.assignments[i].resource.* that were removed.

No behaviour change.

Closes #612.
