+++ packages/app/src/actions/_updateShopping.ts
@@ -1,6 +1,6 @@
 import {CURRENCIES, Scalar, Unit} from "@brewdocs.beer/core";
-import {resourcesOf} from "@/actions/brewableResources";
 import Batch, {ShoppingItem, ShoppingTag} from "@/model/batch";
+import {resourcesOf} from "@/model/brewable";
 import {groupBy, isEqual} from "@/utils/func";
 import {parseNumberString} from "@/utils/math";
