+++ packages/app/src/routes/batches.tsx
@@ -1,10 +1,10 @@
 import {createFileRoute} from "@tanstack/react-router";
 import {useCallback} from "react";
-import {batchProgress} from "@/actions/batchProgress";
 import {Crumb, useBreadcrumbs} from "@/component/breadcrumbs/context";
 import PanelSwitcher from "@/component/panel-switcher";
 import PanelSwitcherContent from "@/component/panel-switcher/content";
 import Batch from "@/model/batch";
+import {batchProgress} from "@/model/batchProgress";
 import {PhaseType} from "@/model/brewable";
 import BrewList from "@/screen/batch-list";
